♻️ refactor: 기존 라디오 버튼 선택 데이터인 성별과 근무 타입 enum화, 무한 참조 에러 해결 #434
- 기존 RadioGroup에서 선택을 관리하던 데이터인 성별과 근무타입을 enum 기반으로 전환
- 이에 따른 /utils/transformers.ts와 /constants/post.ts간 순환 참조 문제 해결 위해 formFields 관련 변수들을 formFields.ts로 분리
- 기존 사용처에서의 import 문 수정
