+++ src/components/WriteDocuments/LaborContractWriteForm.tsx
@@ -1,8 +1,7 @@
 import {
-  initialLaborContractEmployeeInfo,
   LaborContractFormField,
   LaborContractFormFields,
-} from '@/constants/documents';
+} from '@/constants/formFields';
 import {
   DocumentType,
   LaborContractDataResponse,
@@ -23,6 +22,7 @@ import { FormProvider, useForm } from 'react-hook-form';
 import ValidatedSubmitButton from '@/components/Document/write/ValidatedSubmitButton';
 import { renderField } from '@/components/Document/write/renderField';
 import Button from '../Common/Button';
+import { initialLaborContractEmployeeInfo } from '@/constants/documents';
 
 // 필수 검증 필드 목록
 const REQUIRED_FIELDS: Array<keyof LaborContractEmployeeInfo | 'phone'> = [
